pyproject.toml
[project]
name = "risklens"
version = "0.1.0"
description = "RiskLens: block-level climate exposure visualizer for LA"
requires-python = ">=3.11"
dependencies = [
  # Web
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "tenacity>=9.0",

  # Geo
  "geopandas>=1.0",
  "shapely>=2.0",
  "pyproj>=3.6",
  "rasterio>=1.3",
  "rioxarray>=0.17",
  "xarray>=2024.7",
  "netcdf4>=1.7",
  "zarr>=2.18,<3.0",
  "dask[array]>=2024.7",
  "fsspec>=2024.6",
  "s3fs>=2024.6",
  "h3>=4.1",
  "pyarrow>=17.0",

  # Storage
  "duckdb>=1.1",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "geoalchemy2>=0.15",
  "alembic>=1.13",
  "redis>=5.0",

  # Numerics
  "numpy>=1.26",
  "pandas>=2.2",
  "scipy>=1.13",

  # LLM (optional, picked at runtime via LLM_PROVIDER)
  "anthropic>=0.34",
  "openai>=1.40",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "ruff>=0.6",
  "mypy>=1.11",
  "ipykernel>=6.29",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
